# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insta API",
        "description": "API документация для проекта Insta",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/posts/{post}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comments for a post",
                "description": "Returns a list of comments for a specific post",
                "operationId": "18034e7591f147f0c74199972bffba77",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "https://yourdomain.com/storage/avatars/avatar.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "comment": {
                                                "type": "string",
                                                "example": "Nice post!"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "28/04/2025 14:30"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add a comment to a post",
                "description": "Adds a new comment to a specific post",
                "operationId": "5a32d72d0593b3bec9bec52d3dcb988d",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "This is a new comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "https://yourdomain.com/storage/avatars/avatar.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "comment": {
                                            "type": "string",
                                            "example": "This is a new comment"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "28/04/2025 14:30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/posts/{post}/comments/{comment}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update a comment",
                "description": "Updates an existing comment on a post",
                "operationId": "95c598bfa38bd2d6701ce36ad53e195b",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Updated comment text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "https://yourdomain.com/storage/avatars/avatar.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "comment": {
                                            "type": "string",
                                            "example": "Updated comment text"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "28/04/2025 14:30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment",
                "description": "Deletes a comment from a post",
                "operationId": "0eb7d2b813b46cdf69924babc0216189",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Paste your Sanctum access token here",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}